Removed all programming related settings
As with Emacs I end up spending an amount of time which just not
justifiable, to configure Vim as a programming environment. Vim may be a
convenient tool to edit text in a clever way, but I do not think that
cleverness will really save me time on a day-to-day basis compared to
say CLion.

I give up for this time.