Better default makeprg
I was about to dump the vim-cmake plugin I am using, because it does not
set makeprg if I don't run CMake, which forces me to rebuild everything,
but changing the makeprg default value makes things much better of
course. It feels like the vim-cmake plugin somehow should fix this for
me, but my new makeprg default value is quite sane, so let's keep it
simple. I like the fact that I pretty much understand everything in the
plugin code, because it is fairly small and straightforward.