More mappings for jumping around
According to the manual, they should work with both "'" and "`", so
using "'" seems better on a Swedish keyboard. Then the second character
is sometimes also difficult to access on a Swedish keyboard. But not
even all the easy combinations seem to work (like "'<" seems to work,
but not "'>"). This commit should solve all cases. Finally, "']" and
"'[" seem to jump to unexpected places (like the beginning and the end
of the buffer. Oh well...