Minor adjustments
My idea was to go through my Vim configuration to identify potential for
improvements, but more I look at it, the more I think it fits my needs
very well, and it's not particularly unclean. It's not very large
either.

Among others, I do not think switching to NeoVim, or even investigating
it, is worth it. Some reading I have done seems to confirm that the
grass is usually not greener on the other side. Telescope, for instance,
seems to provide worse performance than FZF. FZF, on the other hand, is
one of the most amazing tools I have ever used, so why change?

When it comes to Treesitter, it does not seem like just milk and honey
either. The CPU cost seems significant, and not all language grammars
seem to have good quality. The solution I have today is not that bad, I
never find myself lacking better syntax highlighting.

Treesitter could potentially provide cleverer selections/navigation too,
but I'm not even using the full extent of what Vim has to offer in that
area, so I intend to learn that better, before even thinking about
looking for something better.

All-in-all, the Vim ecosystem is really amazing.